requirements.txt
requests
pandas
math
flatten_json
numpy
datetime
google-cloud-storage
google-cloud-secret-manager
google-cloud-core
google-auth
json